v1.3.0 IssueID #3002: docker IssueID #956: flux IssueID #3140: vista IssueID #3018: v1.3.0
Added:
docs/flux.rst
docs/releases/1_3_0.rst
docs/skyline.flux.rst
docs/skyline.vista.rst
docs/vista.rst
updates/sql/v1.3.0.sql

Modified:
dev-requirements.txt
docs/alert-testing.rst
docs/boundary.rst
docs/conf.py
docs/crucible.rst
docs/development/building-documentation.rst
docs/development/dawn.rst
docs/development/index.rst
docs/getting-started.rst
docs/index.rst
docs/installation.rst
docs/ionosphere_echo.rst
docs/mirage.rst
docs/panorama.rst
docs/rebrow.rst
docs/releases.rst
docs/requirements.rst
docs/roadmap.rst
docs/running-in-python-virtualenv.rst
docs/running-multiple-skylines.rst
docs/skyline.rst
docs/skyline.webapp.rst
requirements.txt
skyline/analyzer/analyzer.py
skyline/analyzer_dev/agent.py
skyline/skyline_version.py
skyline/tsfresh_features/scripts/pytz.rst
utils/dawn/skyline.dawn.sh
utils/dawn/skyline.docker.dawn.sh
utils/docker/configure.sh
utils/docker/init.sh

Deleted:
skyline/docker.settings.py

+++ utils/docker/configure.sh
@@ -96,39 +96,47 @@ if [ -f /opt/skyline/github/skyline/skyline/docker.settings.py ]; then
   echo "Populated variables in the Skyline settings.py using /opt/skyline/github/skyline/skyline/docker.settings.py"
   sleep 1
 else
-  if [ ! -f /opt/skyline/github/skyline/skyline/settings.py.original ]; then
-    cat /opt/skyline/github/skyline/skyline/settings.py > /opt/skyline/github/skyline/skyline/settings.py.original
-  fi
-  # @modified 20180823 - Bug #2552: seed_data.py testing with UDP does not work (GH77)
-  # Only requires a public IP if Grpahite is going to pickle to it, but seeing as
-  # this is a test node, make it 127.0.0.1 as there are no iptables on the IP or
-  # ports 2025 or 2024
-  #    | sed -e "s/HORIZON_IP = .*/HORIZON_IP = '$YOUR_SERVER_IP_ADDRESS'/g" \
-  cat /opt/skyline/github/skyline/skyline/settings.py.original \
-    | sed -e "s/SERVER_METRICS_NAME = .*/SERVER_METRICS_NAME = 'skyline-docker-skyline-1'/g" \
-    | sed -e "s/REDIS_PASSWORD = .*/REDIS_PASSWORD = '$REDIS_PASSWORD'/g" \
-    | sed -e 's/PANORAMA_ENABLED = .*/PANORAMA_ENABLED = True/g' \
-    | sed -e "s/WEBAPP_AUTH_USER = .*/WEBAPP_AUTH_USER = '$WEBAPP_AUTH_USER'/g" \
-    | sed -e 's/PANORAMA_ENABLED = .*/PANORAMA_ENABLED = True/g' \
-    | sed -e "s/WEBAPP_AUTH_USER_PASSWORD = .*/WEBAPP_AUTH_USER_PASSWORD = '$WEBAPP_AUTH_USER_PASSWORD'/g" \
-    | sed -e "s/WEBAPP_ALLOWED_IPS = .*/WEBAPP_ALLOWED_IPS = ['127.0.0.1', '$YOUR_OTHER_IP_ADDRESS']/g" \
-    | sed -e "s/SKYLINE_URL = .*/SKYLINE_URL = 'https:\/\/$YOUR_SKYLINE_SERVER_FQDN'/g" \
-    | sed -e 's/MEMCACHE_ENABLED = .*/MEMCACHE_ENABLED = True/g' \
-    | sed -e "s/PANORAMA_DBUSER = .*/PANORAMA_DBUSER = 'root'/g" \
-    | sed -e "s/PANORAMA_DBHOST = .*/PANORAMA_DBHOST = 'mysql'/g" \
-    | sed -e "s/REDIS_SOCKET_PATH = .*/REDIS_SOCKET_PATH = '\/tmp\/docker\/redis.sock'/g" \
-    | sed -e "s/HORIZON_IP = .*/HORIZON_IP = '$YOUR_SERVER_IP_ADDRESS'/g" \
-    | sed -e "s/MEMCACHE_ENABLED = .*/MEMCACHE_ENABLED = True/g" \
-    | sed -e "s/CARBON_HOST = .*/CARBON_HOST = 'skyline-docker-graphite-statsd-1'/g" \
-    | sed -e "s/PANORAMA_DBUSERPASS = .*/PANORAMA_DBUSERPASS = '$MYSQL_SKYLINE_PASSWORD'/g" > /opt/skyline/github/skyline/skyline/settings.py
-  if [ $? -ne 0 ]; then
-    echo "error :: failed to populate the variables in /opt/skyline/github/skyline/skyline/settings.py"
-    exit 1
-  else
-    echo "Populated variables in the Skyline settings.py"
+  if [ -f /opt/skyline/github/skyline/skyline/docker.settings.py.default ]; then
+    if [ ! -f /opt/skyline/github/skyline/skyline/settings.py.original ]; then
+      cat /opt/skyline/github/skyline/skyline/settings.py > /opt/skyline/github/skyline/skyline/settings.py.original
+    fi
+    cat /opt/skyline/github/skyline/skyline/docker.settings.py.default > /opt/skyline/github/skyline/skyline/settings.py
+    echo "Populated variables in the Skyline settings.py using /opt/skyline/github/skyline/skyline/docker.settings.py"
     sleep 1
+  else
+    if [ ! -f /opt/skyline/github/skyline/skyline/settings.py.original ]; then
+      cat /opt/skyline/github/skyline/skyline/settings.py > /opt/skyline/github/skyline/skyline/settings.py.original
+    fi
+    # @modified 20180823 - Bug #2552: seed_data.py testing with UDP does not work (GH77)
+    # Only requires a public IP if Grpahite is going to pickle to it, but seeing as
+    # this is a test node, make it 127.0.0.1 as there are no iptables on the IP or
+    # ports 2025 or 2024
+    #    | sed -e "s/HORIZON_IP = .*/HORIZON_IP = '$YOUR_SERVER_IP_ADDRESS'/g" \
+    cat /opt/skyline/github/skyline/skyline/settings.py.original \
+      | sed -e "s/SERVER_METRICS_NAME = .*/SERVER_METRICS_NAME = 'skyline-docker-skyline-1'/g" \
+      | sed -e "s/REDIS_PASSWORD = .*/REDIS_PASSWORD = '$REDIS_PASSWORD'/g" \
+      | sed -e 's/PANORAMA_ENABLED = .*/PANORAMA_ENABLED = True/g' \
+      | sed -e "s/WEBAPP_AUTH_USER = .*/WEBAPP_AUTH_USER = '$WEBAPP_AUTH_USER'/g" \
+      | sed -e 's/PANORAMA_ENABLED = .*/PANORAMA_ENABLED = True/g' \
+      | sed -e "s/WEBAPP_AUTH_USER_PASSWORD = .*/WEBAPP_AUTH_USER_PASSWORD = '$WEBAPP_AUTH_USER_PASSWORD'/g" \
+      | sed -e "s/WEBAPP_ALLOWED_IPS = .*/WEBAPP_ALLOWED_IPS = ['127.0.0.1', '$YOUR_OTHER_IP_ADDRESS']/g" \
+      | sed -e "s/SKYLINE_URL = .*/SKYLINE_URL = 'https:\/\/$YOUR_SKYLINE_SERVER_FQDN'/g" \
+      | sed -e 's/MEMCACHE_ENABLED = .*/MEMCACHE_ENABLED = True/g' \
+      | sed -e "s/PANORAMA_DBUSER = .*/PANORAMA_DBUSER = 'root'/g" \
+      | sed -e "s/PANORAMA_DBHOST = .*/PANORAMA_DBHOST = 'mysql'/g" \
+      | sed -e "s/REDIS_SOCKET_PATH = .*/REDIS_SOCKET_PATH = '\/tmp\/docker\/redis.sock'/g" \
+      | sed -e "s/HORIZON_IP = .*/HORIZON_IP = '$YOUR_SERVER_IP_ADDRESS'/g" \
+      | sed -e "s/MEMCACHE_ENABLED = .*/MEMCACHE_ENABLED = True/g" \
+      | sed -e "s/CARBON_HOST = .*/CARBON_HOST = 'skyline-docker-graphite-statsd-1'/g" \
+      | sed -e "s/PANORAMA_DBUSERPASS = .*/PANORAMA_DBUSERPASS = '$MYSQL_SKYLINE_PASSWORD'/g" > /opt/skyline/github/skyline/skyline/settings.py
+    if [ $? -ne 0 ]; then
+      echo "error :: failed to populate the variables in /opt/skyline/github/skyline/skyline/settings.py"
+      exit 1
+    else
+      echo "Populated variables in the Skyline settings.py"
+      sleep 1
+    fi
   fi
-fi
 
 if [ -f /tmp/docker_build ]; then
   echo "Not installing skyline DB during docker build"
